+++ flocx_ui/content/flocx/panel.py
@@ -2,5 +2,6 @@ from django.utils.translation import ugettext_lazy as _
 import horizon
 
 class Flocx(horizon.Panel):
-  name = _("Flocx")
-  slug = "flocx"
+    name = _("Flocx")
+    slug = "flocx"
+