+++ flocx_ui/content/flocx/views.py
@@ -1,4 +1,5 @@
 from django.views import generic
 
 class IndexView(generic.TemplateView):
-  template_name = 'project/flocx/index.html'
+    template_name = 'project/flocx/index.html'
+