+++ flocx_ui/static/dashboard/project/flocx/offer-list/offer-list.controller.js
@@ -5,15 +5,34 @@
     .module('horizon.dashboard.project.flocx', [])
     .controller('FlocxOfferListController', FlocxOfferListController);
 
-  FlocxOfferListController.$inject = [ '$http' ];
+  FlocxOfferListController.$inject = [
+    'horizon.app.core.openstack-service-api.flocx'
+  ];
 
-  function FlocxOfferListController(_$http) {
+  function FlocxOfferListController(flocx) {
     var ctrl = this;
 
-    ctrl.items = [
-      { name: 'abc', id: 123 },
-      { name: 'efg', id: 345 },
-      { name: 'hij', id: 678 }
-    ];
+    ctrl.offers = [];
+
+    init();
+
+    function init () {
+      retrieveOffers();
+    }
+
+    function retrieveOffers () {
+      // Get the list of offers from the flocx-market
+      flocx.getOffers().then(onGetOffers);
+    }
+
+    function onGetOffers (offers) {
+      ctrl.offers = offers.map(function (offer) {
+        var startTime = offer.start_time.slice(1, -1); // get rid of parenthesis
+        var endTime = offer.end_time.slice(1, -1); // get rid of parenthesis
+        offer.start_time = new Date(startTime).toLocaleString();
+        offer.end_time = new Date(endTime).toLocaleString();
+        return offer;
+      });
+    }
   }
 })();